+++ proofs/src/poly/kzg/mod.rs
@@ -11,6 +11,9 @@ use std::marker::PhantomData;
 
 use midnight_curves::pairing::Engine;
 
+#[cfg(feature = "fewer-point-sets")]
+use super::query::Query;
+
 /// Multiscalar multiplication engines
 pub mod msm;
 /// KZG commitment scheme
@@ -23,6 +26,8 @@ use ff::Field;
 use group::Group;
 use midnight_curves::pairing::MultiMillerLoop;
 use rand_core::OsRng;
+#[cfg(feature = "fewer-point-sets")]
+pub use utils::compute_dummy_queries;
 
 #[cfg(feature = "truncated-challenges")]
 use crate::utils::arithmetic::{truncate, truncated_powers};
@@ -97,7 +102,7 @@ where
 
     fn multi_open<T: Transcript>(
         params: &Self::Parameters,
-        prover_query: &[ProverQuery<E::Fr>],
+        queries: &[ProverQuery<E::Fr>],
         transcript: &mut T,
     ) -> Result<(), Error>
     where
@@ -119,12 +124,27 @@ where
                 .unwrap()
         }
 
+        // Add dummy queries to reduce the number of distinct multi-open point sets.
+        #[cfg(feature = "fewer-point-sets")]
+        let queries = &{
+            let mut queries = queries.to_vec();
+            let pairs: Vec<_> = queries.iter().map(|q| (q.get_commitment(), q.point)).collect();
+            for (idx, point) in compute_dummy_queries(&pairs) {
+                let poly = queries[idx].poly;
+                transcript
+                    .write(&eval_polynomial(poly, point))
+                    .map_err(|_| Error::OpeningError)?;
+                queries.push(ProverQuery::new(point, poly));
+            }
+            queries
+        };
+
         // Refer to the halo2 book for docs:
         // https://zcash.github.io/halo2/design/proving-system/multipoint-opening.html
         let x1: E::Fr = transcript.squeeze_challenge();
         let x2: E::Fr = transcript.squeeze_challenge();
 
-        let (poly_map, point_sets) = construct_intermediate_sets(prover_query)?;
+        let (poly_map, point_sets) = construct_intermediate_sets(queries)?;
 
         let mut q_polys = vec![vec![]; point_sets.len()];
 
@@ -218,19 +238,35 @@ where
     }
 
     fn multi_prepare<'com, T: Transcript>(
-        verifier_query: &[VerifierQuery<'com, E::Fr, KZGCommitmentScheme<E>>],
+        queries: &[VerifierQuery<'com, E::Fr, KZGCommitmentScheme<E>>],
         transcript: &mut T,
     ) -> Result<DualMSM<E>, Error>
     where
         E::Fr: Sampleable<T::Hash> + Ord + Hash + Hashable<T::Hash>,
         E::G1: 'com + Hashable<T::Hash> + CurveExt<ScalarExt = E::Fr>,
     {
+        // Add dummy queries to reduce the number of distinct multi-open point sets.
+        #[cfg(feature = "fewer-point-sets")]
+        let queries = &{
+            let mut queries = queries.to_vec();
+            let pairs: Vec<_> = queries.iter().map(|q| (q.commitment.clone(), q.point)).collect();
+            for (idx, point) in compute_dummy_queries(&pairs) {
+                queries.push(VerifierQuery {
+                    point,
+                    commitment_label: queries[idx].commitment_label.clone(),
+                    commitment: queries[idx].commitment.clone(),
+                    eval: transcript.read().map_err(|_| Error::SamplingError)?,
+                });
+            }
+            queries
+        };
+
         // Refer to the halo2 book for docs:
         // https://zcash.github.io/halo2/design/proving-system/multipoint-opening.html
         let x1: E::Fr = transcript.squeeze_challenge();
         let x2: E::Fr = transcript.squeeze_challenge();
 
-        let (commitment_map, point_sets) = construct_intermediate_sets(verifier_query)?;
+        let (commitment_map, point_sets) = construct_intermediate_sets(queries)?;
 
         let mut q_coms: Vec<_> = vec![vec![]; point_sets.len()];
         let mut q_eval_sets = vec![vec![]; point_sets.len()];